@@ -1,21 +1,6 @@
-[tool.black]
-target-version = ['py310']
-include = '\.pyi?$'
-skip-string-normalization = true
-
-[tool.isort]
-profile = "black"
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
-line_length = 88
-known_first_party = ["business_rules"]
-
 [tool.poetry]
 name = "business-rules"
-version = "1.6.1"
+version = "1.7.0"
 description = "Python DSL for setting up business intelligence rules that can be configured without code [https://github.com/venmo/business-rules]"
 authors = ["venmo <open-source@venmo.com>"]
 readme = "README.md"
@@ -26,15 +11,14 @@ license='MIT'
 test = ["pytest", "pytest-cov"]
 
 [tool.poetry.dependencies]
-python = ">=3.7,<4.0"
-pytz = "^2023.3"
+python = ">=3.9,<4.0"
 
 [tool.poetry.group.dev.dependencies]
-mock = "*"
-nose = "*"
-nose-run-line-number = "*"
-pytest = "*"
-pytest-cov = "*"
+mock = "^5.2.0"
+nose = "^1.3.7"
+nose-run-line-number = "^0.0.2"
+pytest = "^8.4.2"
+pytest-cov = "^7.0.0"
 coverage = "^7.2.3"
 tox = "^4.5.1"
 
